Require select elements to have a value set
On registering UI for shortcode, checks that any attributes of type
"select" have a value set. If not, sets their value to the key of the
first option, as this is the behavior that will be expected by the user
who sees that the first element in the select list appears to be
selected on the form.