Add basic API for registering callbacks on attributes
This adds a basic method for registering a callback function on an
attribute field when registering UI for a shortcode. If the attribute
`callback` is defined on a field and maps to the name of a valid
javascript function, that function will be called on rendering the
shortcode UI and on any changes to that attribute field.

The value of *this* in the callback function will tbe the array of view
models representing all of the attribute input field views, and the
function receives two arguments: the changed value, and the collection
of model attributes.