Remove wp-hooks from browserified output
Originally I tried including wp-hooks as a requireJS module. That didn't
make any sense because it needs to be in the global namespace. So,
removed it.

Also, fixed the namespace in the action call from edit-attribute field
(the namespace registration was updated from `shortcodeUi` to
`shortcake` in the last commit, but editAttributeField was still trying
to call `wp.shortcodeUi.hooks.addAction`)