Un-browserify the attachment field view
I can't see any reason to build in an extra 10K of scripts into this
file which are already being included elsewhere on the page just for the
purpose of encapsulation. I tried to remove all the browserify
dependencies and set this field up as a model for how plugin users could
register their own fields in a more typical Backbone view structure. I
like it a lot better.

I am exposing one additional global, the `iDCache` variable, and that's
somewhat intentional - I'd like for filters on updating attachment to
have access to that data to avoid having to make additional ajax calls
to look it up. The variable name and the way in which it's exposed is
pretty ugly though, and I think it would do better as a public method on
the view...

Also, browserify is still processing this file, I'm just loading the
version from `src` rather than the browserified `build` file. It would
be good to come up with an organizational structure that made more sense
for these view files.