Move editAttributeField into main browserified bundle
Following @jitendraharpalani's comments, it makes sense for this to be
bundled into the main plugin js file and avoid the extra unnecessary
HTTP request.

It's still architected basically as a standalone view, which I think is
a good thing. If we keep the pattern used for composing views fairly
standard and standalone, it will be easier to give examples for
extending these views. Everything needed to extend a view should already
be accessible through extending the `sui.views.editAttributeField`
field.