Fire 'register_shortcode_ui' hook on enqueuing editor
Firing this hook on `init` is too broad and doesn't offer users any
better control over when they register their scripts than if we just
directed them to register them in `init` itself. I think th
`wp_enqueue_editor` action is a better fit, as it happens early enough
to register and enqueue any scripts or styles that will be necessary,
and only fires on pages where a WP_Editor is actually being set up.

This action also gets a supports argument, so it's possible to register
scripts only on editors which include tinymce, for example.