Add "Fetcher" class for bulk fetching on shortcode views
Adds a Fetcher utility class to combine and bulk-send requests for
shortcode previews. Currently - until I can find some reliable events on
MCE processing to hook to - this is being done in a very naive way:
every time a request is made, it sets a half-second timeout, so that any
other requests needed can be gathered and sent together.