Reset the shortcode regexp's index between each match
Because the shortcode Regexp defined by `wp.shortcode.rexexp` is a
global matcher, each time it is `exec()`'d, it preserves the last index
searched up to, and begins searching from that index of the target
string the next time it is run over that target string.

This fixes that problem by resetting the lastIndex counter to 0 before
executing the regex.