Add SCRIPT_DEBUG check back in
I removed the check for SCRIPT_DEBUG a while while trying to
troubleshoot this. Its probably a good thing to respect this constant
and serve the minified script file if requested. This preserves the
switch to enqueueing the full version of select2 4.0.3, while also
serving a minified or non-minified version depending on the SCRIPT_DEBUG
constant.