For select fields, set model value on initial render
Select fields are a bit of a special case as far as attribute fields go,
in that the select element's value in the UI can appear to be set
without the model value itself being defined. This change handles cases
where there's no empty value in the select field, so the first option
should function as a default value.

When a select field is rendered, if the model's value is empty and the
first element isn't an empty option, this will trigger an initial
"change" event to set the model's value to the value of the option which
is visibly selected, to avoid confusion.