Use select2.full.js for compatibility with other plugins which enqueue it (#666)
* compatibility with WooCommerce 2.6.8 select2

* Add SCRIPT_DEBUG check back in

I removed the check for SCRIPT_DEBUG a while while trying to
troubleshoot this. Its probably a good thing to respect this constant
and serve the minified script file if requested. This preserves the
switch to enqueueing the full version of select2 4.0.3, while also
serving a minified or non-minified version depending on the SCRIPT_DEBUG
constant.

* Bump plugin version number for cache-busting

Noticed that this was never updated when 0.7.0 was released. Bumping the
version number here to bust any cached scripts or styles. (It'll be
bumped again by the time this is released in a patch version.)