Add `send_to_editor` filter
Adds a filter which can be used for client-side validation of a
shortcode before sending it to the editor. Basic concept for addressing
issue #615; for discussion only at this point. (I think this needs some
kind of syntactic sugar to help work with it, as the API this exposes
doesn't follow any established conventions in the WP universe and isn't
self-explanatory at all.)