Optionally set checkbox values as 'true' or 'false'
Adds the attribute option 'explicitbool' for checkboxes. If a checkbox
is registered with 'explicitbool' => true, then the shortcode will be
formatted with that attribute string like `attribute="false"` if
unchecked. If the shortcode attributes are retrieved using the shortcode
tag as the third parameter to `shortcode_atts()`, then this value will
be converted back to a boolean automatically.

This allows better handling of default values for attributes -
otherwise, if an attribute was set with a default value of "true", then
even if it was unchecked, the next time the shortcode was opened to
edit, that default would be applied and the checkbox would be reset to
checked.