Re-add deprecated function
Up until 0.7.0, we had a public class method exposed on
editAttributeFieldAttachment called `getFromCache`, which was being used
elsewhere to hook into attachment preview rendering and other useful
attachment needs. (For example, the Image Shortcake plugin was using it
to pull alt, caption, and other values from the attachment to
prepopulate those fields.)

I think it would still be helpful to expose this cache, even though it
isn't needed anymore internally, just because it costs next to nothing
to have in place, and is useful for integration with other fields.