Rename class files following WPCS rules
The classes defining the various fields used in the plugin have gotten a
bin inconsistant in naming conventions, with some of them being named
Shortcake_Field_{FieldType} and others Shortcode_UI_Field_{FieldType}.

I figured that, since recent rules added to WPCS have started flagging
the filenames of these classes for not matching the class name, this was
a good time to go through and standardize on a single naming convention.
Kept the file names the same, moved the classes into the Shortcode_UI
namespace and removed the plugin name from the class names.