Disable the media button if no shortcodes are registered
The "Add post element" button does nothing if no shortcode UIs are
registered. This prevents it from being attached to the editor if no
shortcodes are available to be inserted through it.

(Note that because `register_shortcode_ui` has access to the editor ID
as context, it's possible for shortcodes to be registered in some editor
contexts and not others. This handles that sort of... to the same degree
that `register_shortcode_ui` does, at least.)