Fix select issue (#745)
* Ensure data value is an array before filtering

During work in #707 to improve support for multi-select fields,
handling of comma separated strings was removed as it may have
been confusing if an option value had commas in it.

See 9894162ab8bf02941e9611766d5cc3d63d337e96

After this change, string values for single select elements are
filtered as if they were arrays. This causes only the first
character of a string to be compared to the select value. This in
turn results in all select fields set to first option because
nothing is seen as selected.

We can accept arrays, but also `split()` with no separator to turn
strings into an array so that everything is parsed the same.

* Use _.contains instead of filter/empty

* Split on comma rather than just casting to array