Don't override global select2 z-indexing
Because the select2 fields we're using here need to show up just over
the absolutely positioned media modal, we're setting a z-index value of
just above that modal.

However, there are other places in the WP admin (like the customizer)
which require a much higher z-index value, so it would be better to
target only trhe select2 instance which we're using.

This commit uses the "theme" argument to assign an additional classname
to the select2 container, and then targets the container in CSS using
the new classname.