Use select2 dropdownParent attr for css targetting
When interacting with a select2 field, this attaches the dropdown
container as a child of the attribute field view, using the
`dropdownParent` option in select2.

This allows us to target the specific fields that we need to style with
css using parent selectors, without resorting to hacks like trying to
set a classname on the container element when it's created.