Avoid saving "null" values on empty select2 fields
We're coercing the value of select2 fields to a string, in order to
handle some edge cases in post and term select fields with multiple=true.

However, after deleting all values from a multiple select field which
had been previously updated with one or more values, that attribute's
value is `null`.

We don't want to coerce this empty value to the string "null".
This PR fixes the issue by replacing null attribute values with the
empty string, which doesn't get saved when the model is run through
`Shortcode.formatShortcode()`.