Remove unused method for fetching preview of single shortcode
Removes the unused "fetch()" method from Shortcode View constructor (it
was confusingly making a request to the 'do_shortcode' ajax endpoint,
which was removed a long time ago). Also removes the nonce which is set
with each separate query in the bulk_do_shortcode request - since #743,
a single nonce is being sent as a top-level field on the request and is
being checked by the 'bulk_do_shortcode' endpoint.