Lock wpcs at latest release (0.14.0)
dev-master on WPCS is starting to rename some sniffs that are causing
warnings in our ruleset. This change locks the WPCS package to the
latest release instead of running amster branch, updates PHP
Codensniffer to the latest release, and rolls back the changes to the
ruleset I tried in the last commit to keep up with the WPCS rule
changes.