Avoid outputting duplicate underscore templates
Many of the attribute fields in this plugin were outputting their
templates on the `shortcode_ui_loaded_editor` hook, which fires every
time an editor is created. As a result, duplicate templates will be
created with identical IDs.

This addresses that problem by moving the template enqueues to the
`enqueue_shortcode_ui` hook, which only fires once per page load, and
from that hook, adding the template output functions onto the
admin_print_footer_scripts hook, so that they show up in a reasonable
place rather than in the middle of the page markup.